Refactoring: Extract VerificationAnalysis to its own file
