+++ core/src/main/scala/stainless/verification/VerificationComponent.scala
@@ -27,25 +27,6 @@ object VerificationComponent extends SimpleComponent {
 
   implicit val debugSection = DebugSectionVerification
 
-  trait VerificationAnalysis extends AbstractAnalysis {
-    val program: Program { val trees: stainless.trees.type }
-    val results: Map[VC[program.trees.type], VCResult[program.Model]]
-
-    lazy val vrs: Seq[(VC[stainless.trees.type], VCResult[program.Model])] =
-      results.toSeq.sortBy { case (vc, _) => (vc.fd.name, vc.kind.toString) }
-
-    override val name = VerificationComponent.name
-
-    override type Report = VerificationReport
-
-    override def toReport = new VerificationReport(vrs map { case (vc, vr) =>
-      val time = vr.time.getOrElse(0L) // TODO make time mandatory (?)
-      val status = VerificationReport.Status(vr.status)
-      val solverName = vr.solver map { _.name }
-      VerificationReport.Record(vc.fd, vc.getPos, time, status, solverName, vc.kind.name)
-    })
-  }
-
   private def check(funs: Seq[Identifier], p: StainlessProgram, ctx: inox.Context): Map[VC[p.trees.type], VCResult[p.Model]] = {
     val injector = AssertionInjector(p, ctx)
     val encoder = inox.ast.ProgramEncoder(p)(injector)
